Fix fatal PHP 7.1 error with sites where the $actions pulls as a string instead of an array
PHP 7.1 has a fatal error when `$actions[] = "<a href='$settings_page_url'>" . __( 'Settings', 'simple-history' ) . '</a>';` is executed when it's been handed $actions being a string instead of an array. I'm assuming PHP 7.1 doesn't do this automatic string to array conversion anymore due to performance reasons.

I would love to see this officially adopted in a future version of the plugin as I'm thinking others out there might encounter this issue after updating their hosting environment to PHP 7.1+. As an aside, I'm thinking this might only trigger an issue on WordPress sites which have been active for a few years since it's an older site (though running WP 4.9) that had an issue while I have other similar sites on the same setup without this issue. Either way, it'd be good to accommodate for this potential fatal error.

Thanks for the great plugin!